Add screenshots for plugin type descriptions
Two screenshots showcasing functionality of plugin type descriptions feature. The extensions admin ui image shows that the extension description is still used when the context of the extension is in general (i.e. not plugin type specific), whereas the workflows event selector shows that we update the description when in the context of a specific plugin type.