Add screenshots for plugin type descriptions (#25)
* Add screenshots for plugin type descriptions

Two screenshots showcasing functionality of plugin type descriptions feature. The extensions admin ui image shows that the extension description is still used when the context of the extension is in general (i.e. not plugin type specific), whereas the workflows event selector shows that we update the description when in the context of a specific plugin type.

* Delete plugin_type_descriptions_general_eaui.png

* Delete plugin_type_descriptions_workflows_event.png

* Use more zoomed in images for plugin type descriptions